requirements.txt
alembic==1.13.1
aniso8601==9.0.1
annotated-types==0.6.0
attrs==23.2.0
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
Deprecated==1.2.14
Flask==3.0.2
Flask-Limiter==3.5.1
Flask-Mail==0.9.1
Flask-Migrate==4.0.5
flask-restx==1.3.0
Flask-SQLAlchemy==3.1.1
idna==3.6
importlib-metadata==7.0.1
importlib-resources==6.1.1
itsdangerous==2.1.2
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
limits==3.9.0
Mako==1.3.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
ordered-set==4.1.0
packaging==23.2
passlib==1.7.4
psycopg2==2.9.9
psycopg2-binary==2.9.9
pydantic==2.6.1
pydantic-settings==2.1.0
pydantic_core==2.16.2
Pygments==2.17.2
PyJWT==2.8.0
python-dotenv==1.0.1
pytz==2024.1
referencing==0.33.0
requests==2.31.0
rich==13.7.0
rpds-py==0.17.1
SQLAlchemy==2.0.25
typing_extensions==4.9.0
urllib3==2.2.1
Werkzeug==3.0.1
wrapt==1.16.0
zipp==3.17.0
